% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_PC.R
\name{align_PC}
\alias{align_PC}
\title{align_PC}
\usage{
align_PC(
  PC,
  ia = 100,
  add = "all",
  isocount = 2,
  plan = "multiprocess",
  pick_best = "rt_match"
)
}
\arguments{
\item{pick_best}{}
}
\value{

}
\description{
align_PC
}
