% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plots.R
\name{as.sunburstDF}
\alias{as.sunburstDF}
\title{as.sunburstDF}
\usage{
as.sunburstDF(DF, valueCol = NULL)
}
\arguments{
\item{valueCol}{}
}
\value{

}
\description{
This function was written by ismirsehregal on stackoverflow https://stackoverflow.com/questions/57395424/how-to-format-data-for-plotly-sunburst-diagram
}
