% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_peaks.R
\name{assemble_peaks}
\alias{assemble_peaks}
\title{assemble_peaks}
\usage{
assemble_peaks(peak.grp, s, e, breakP)
}
\arguments{
\item{peak.grp}{peak.grp}

\item{s}{start}

\item{e}{end}

\item{breakP}{break point}
}
\description{
assemble_peaks
}
\keyword{internal}
