% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_alignment.R
\name{assess_alignment}
\alias{assess_alignment}
\title{assess_alignment}
\usage{
assess_alignment(Matches_BM_NPPpeaks, Unmatched_BM_NPPpeaks, g_table)
}
\arguments{
\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{Unmatched_BM_NPPpeaks}{Unmatched_BM_NPPpeaks}

\item{g_table}{g_table}
}
\description{
assess_alignment
}
\keyword{internal}
