% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_functions.R
\name{assign_groupID_column}
\alias{assign_groupID_column}
\title{assign_groupID_column}
\usage{
assign_groupID_column(dt, column_name, by)
}
\arguments{
\item{by}{}
}
\description{
assign_groupID_column
}
