% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_pick_main_feature.R
\name{best_feature_per_comparison}
\alias{best_feature_per_comparison}
\title{best_feature_per_comparison}
\usage{
best_feature_per_comparison(dt)
}
\arguments{
\item{dt}{dt}
}
\description{
best_feature_per_comparison
}
\keyword{internal}
