% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_benchmark.R
\name{check_benchmark_input}
\alias{check_benchmark_input}
\title{check_benchmark_input}
\usage{
check_benchmark_input(file, options_path = "generate", from_csv = TRUE, algo)
}
\arguments{
\item{file}{output of \code{\link{find_bench_peaks}}. Can be path to csv file or a data table  object (meaning that is.data.table(file) returns TRUE).}

\item{options_path}{can be a string "generate" in order to use default column names for chosen algo. In the future we might include a possibility to allow the user to choose column names.}

\item{from_csv}{TRUE or FALSE depending on file being a data.table object or a path to a csv}

\item{algo}{tool output format to compare the benchmark against. can be XCMS, XCMS3, El-Maven, OpenMS, MS-DIAL, CompoundDiscoverer or MZmine 2 Outputs from different tools can also be used as long as they are reformatted to one of those types.}
}
\value{
returns a list including the benchmark in a format readable by \code{\link{compare_peaks}}.
}
\description{
Checks the benchmark dataset and brings it into a format readable by \code{\link{compare_peaks}}. All molecules for which the most abundant isotopolgue is not present,
or less than 2 isotopologues are present are deleted. Moreover, isotopologues which appear in only one file are deleted.
}
