% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_IR_biases.R
\name{check_IR_biases}
\alias{check_IR_biases}
\title{check_IR_biases}
\usage{
check_IR_biases(
  Matches_BM_NPPpeaks,
  Matches_BM_NPPpeaks_NPPfeatures,
  g_table,
  b_table
)
}
\arguments{
\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{Matches_BM_NPPpeaks_NPPfeatures}{Matches_BM_NPPpeaks_NPPfeatures}

\item{g_table}{g_table}

\item{b_table}{b_table}
}
\description{
check_IR_biases
}
\keyword{internal}
