% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_peaks.R
\name{check_missing_peaks}
\alias{check_missing_peaks}
\title{check_missing_peaks}
\usage{
check_missing_peaks(
  Matches_BM_NPPpeaks,
  Unmatched_BM_NPPpeaks,
  Matches_BM_NPPpeaks_NPPfeatures,
  g_table
)
}
\arguments{
\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{Unmatched_BM_NPPpeaks}{Unmatched_BM_NPPpeaks}

\item{Matches_BM_NPPpeaks_NPPfeatures}{Matches_BM_NPPpeaks_NPPfeatures}

\item{g_table}{g_table}
}
\description{
check_missing_peaks
}
\keyword{internal}
