% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_pick_algo.R
\name{check_nonTargeted_input}
\alias{check_nonTargeted_input}
\title{check_nonTargeted_input}
\usage{
check_nonTargeted_input(
  ug_table_path,
  g_table_path,
  options_table = "generate",
  algo
)
}
\arguments{
\item{ug_table_path}{path to unaligned table(s)}

\item{g_table_path}{path to aligned table}

\item{options_table}{output from \code{\link{check_benchmark_input}}}

\item{algo}{tool output format of ug_table and g_table. can be XCMS, XCMS3, El-Maven, OpenMS, msDial, CompoundDiscoverer or mzMine. Outputs from different tools can also be used as long as they are reformatted to one of those types.}
}
\value{
returns unaligned and aligned outputs from non-targeted tool in a format readable via \code{\link{compare_peaks}}
}
\description{
check_nonTargeted_input
}
