% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choose_main_peak.R
\name{choose_main_peak}
\alias{choose_main_peak}
\title{choose_main_peak}
\usage{
choose_main_peak(comp_id_b, comp_id_ug, isoabb_b, peak_area_ug,
  peak_height_ug, peak_height_b)
}
\arguments{
\item{peak_height_b}{}
}
\description{
choose_main_peak
}
