% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_classify_false_negative.R
\name{classify_false_negative}
\alias{classify_false_negative}
\title{classify_false_negative}
\usage{
classify_false_negative(dt)
}
\arguments{
\item{dt}{}
}
\value{

}
\description{
classify_false_negative
}
