% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_peak_compare.R
\name{compare_peaks_ug_g}
\alias{compare_peaks_ug_g}
\title{compare_peaks_ug_g}
\usage{
compare_peaks_ug_g(b_table, ug_table, g_table, algo, main_feature_method)
}
\arguments{
\item{algo}{}
}
\description{
compare_peaks_ug_g
}
