% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_peak_compare_no_rt_start.R
\name{compare_peaks_ug_g_no_rt_start}
\alias{compare_peaks_ug_g_no_rt_start}
\title{compare_peaks_ug_g_no_rt_start}
\usage{
compare_peaks_ug_g_no_rt_start(b_table, ug_table, g_table, algo)
}
\arguments{
\item{algo}{}
}
\description{
compare_peaks_ug_g_no_rt_start
}
