% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_funktionen_fuer_alignment.R
\name{count_alignment_errors}
\alias{count_alignment_errors}
\title{count_alignment_errors}
\usage{
count_alignment_errors(DT, main_UTgroups, method = "both")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
count alignment errors occuring within a compound
}
