% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_count_max_error.R, R/comp_functions.R
\name{count_errors_max}
\alias{count_errors_max}
\title{count_errors_max}
\usage{
count_errors_max(dt)

count_errors_max(dt)
}
\arguments{
\item{dt}{}

\item{dt}{}
}
\description{
count_errors_max

count_errors_max
}
