% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_generate_results_text.R
\name{count_split_features}
\alias{count_split_features}
\title{count_split_features}
\usage{
count_split_features(feature_id_ug)
}
\arguments{
\item{feature_id_ug}{}
}
\description{
count_split_features
}
