% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutout_peak.R
\name{cutout_peaks}
\alias{cutout_peaks}
\title{cutout_peaks}
\usage{
cutout_peaks(
  int,
  rt,
  Min.PpP = 10,
  peak.spotting.factor. = 0.01,
  Integration_baseL_factor. = 0.1,
  l = 1,
  r = length(int),
  M0.grp = NA,
  main_adduct.grp = NA,
  Min.Res. = 70
)
}
\arguments{
\item{int}{int}

\item{rt}{rt}

\item{Min.PpP}{Min.PpP}

\item{peak.spotting.factor.}{peak.spotting.factor.}

\item{Integration_baseL_factor.}{Integration_baseL_factor.}

\item{l}{l}

\item{r}{r}

\item{M0.grp}{M0.grp}

\item{main_adduct.grp}{main_adduct.grp}

\item{Min.Res.}{Min.Res.}
}
\description{
detects peaks
}
\keyword{internal}
