% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_choose_main_peak.R
\name{eliminate_duplicates_no_for}
\alias{eliminate_duplicates_no_for}
\title{eliminate_duplicates_no_for}
\usage{
eliminate_duplicates_no_for(comp_id_b, comp_id_ug, isoabb_b, peak_area_ug,
  peak_height_ug, peak_height_b)
}
\arguments{
\item{peak_height_b}{}
}
\description{
eliminate_duplicates_no_for
}
