% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_feature_compare.R
\name{feature_compare}
\alias{feature_compare}
\title{feature_compare}
\usage{
feature_compare(b_table, g_table)
}
\arguments{
\item{b_table}{b_table}

\item{g_table}{g_table}
}
\description{
feature_compare
}
\keyword{internal}
