% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_r_s_error.R
\name{File_con_test}
\alias{File_con_test}
\title{File_con_test}
\usage{
File_con_test(FileName, feature_id)
}
\arguments{
\item{FileName}{FileName}

\item{feature_id}{feature_id}
}
\value{
connected file grps
}
\description{
File_con_test
}
\keyword{internal}
