% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Background.R
\name{Filter_Background}
\alias{Filter_Background}
\title{Filter_Background}
\usage{
Filter_Background(
  PeakCandidates,
  IonSpeciesIdentifier,
  SampleIdentifier,
  Blank_label,
  Factor = 5
)
}
\arguments{
\item{PeakCandidates}{x}

\item{IonSpeciesIdentifier}{Should uniquly identify the ion species (one column name or a vector of column names)}

\item{SampleIdentifier}{column name or vector of column names which uniquely identify measured samples}

\item{Blank_label}{string by which blanks are labeled within the SampleIdentifier}

\item{Factor}{factor by which sample peaks have to outreach blank peaks (mean of blank peaks over all replicates is used if multiple replicates are available)}
}
\value{
PeakCandidates table with additional logical column: Background (TRUE/FALSE)
}
\description{
Takes a PeakCandidates table and lables peaks in non-blank-samples if they do not outreach peaks found in blanks by a certain factor (in area as well as in height).
}
\examples{


}
