% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_WrongCompounds.R
\name{Filter_WrongCompounds}
\alias{Filter_WrongCompounds}
\title{Filter_WrongCompounds}
\usage{
Filter_WrongCompounds(PeakCandidates, ErrorRel_H_Mp1_tol = 10)
}
\arguments{
\item{IsoAbb_col}{}
}
\description{
Filter_WrongCompounds
}
