% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_WrongIsoOrder.R
\name{Filter_WrongIsoOrder}
\alias{Filter_WrongIsoOrder}
\title{Filter_WrongIsoOrder}
\usage{
Filter_WrongIsoOrder(PeakCandidates, SampleIdentifier, MoleculeIdentifier,
  AdductIdentifier, IsotoplogeIdentifier)
}
\arguments{
\item{PeakCandidates}{x}

\item{SampleIdentifier}{column name or vector of column names which uniquely identify measured samples}

\item{MoleculeIdentifier}{column name or vector of column names which uniquely identify sample molecules}

\item{AdductIdentifier}{column name or vector of column names which uniquely identify molecule adducts}

\item{IsotoplogeIdentifier}{column name with numeric representation of theoretical isotopologue abundance}
}
\value{
PeakCandidates table with additional logical column: WrongIsoOrder (TRUE/FALSE)
}
\description{
Takes a PeakCandidates table and lables peaks if their abundance or elution profile does not correspond to the order indicated by the theoretical isotopologue pattern.
}
