% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_best_feature_feature.R
\name{find_best_feature_feature}
\alias{find_best_feature_feature}
\title{find_best_feature_feature}
\usage{
find_best_feature_feature(dt, bys)
}
\arguments{
\item{bys}{}
}
\value{

}
\description{
find_best_feature_feature
}
