% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_main_feature.R
\name{find_best_matching_iso_pairs}
\alias{find_best_matching_iso_pairs}
\title{find_best_matching_iso_pairs}
\usage{
find_best_matching_iso_pairs(dt, iso_1, iso_2, max_possible_found_peaks,
  nf_g_table, factor, iso_1_feature = NA)
}
\arguments{
\item{iso_1_feature}{}
}
\description{
find_best_matching_iso_pairs
}
