% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_main_feature.R
\name{find_main_feature_1}
\alias{find_main_feature_1}
\title{find_main_feature_1}
\usage{
find_main_feature_1(c_table, setting, nf_g_table, nf_b_table)
}
\arguments{
\item{nf_b_table}{}
}
\description{
find_main_feature_1
}
