% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_main_feature.R
\name{find_most_occuring_feature}
\alias{find_most_occuring_feature}
\title{find_most_occuring_feature}
\usage{
find_most_occuring_feature(feature_id_g, peak_area_g)
}
\arguments{
\item{peak_area_g}{}
}
\description{
find_most_occuring_feature
}
