% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_r_s_error.R
\name{find_r_s_error}
\alias{find_r_s_error}
\title{find_r_s_error}
\usage{
find_r_s_error(peak_area_b, peak_area_ug, peak_height_b)
}
\arguments{
\item{peak_height_b}{}
}
\value{

}
\description{
find_r_s_error
}
