% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_generate_options.R
\name{generate_options}
\alias{generate_options}
\title{generate_options}
\usage{
generate_options(raw_b_table, algo)
}
\arguments{
\item{raw_b_table}{raw_b_table}

\item{algo}{algo}
}
\description{
generate_options
}
\keyword{internal}
