% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_generate_results_text.R
\name{generate_results_text}
\alias{generate_results_text}
\title{generate_results_text}
\usage{
generate_results_text(comparison_data)
}
\arguments{
\item{comparison}{output from \code{\link{compare_peaks}}}
}
\value{
returns a list containing containing different performance metrics of non-targeted data pre-processing.
}
\description{
in order to estimate confidence intervals boots trapping is performed on benchmark molecules with R=1000 using \code{\link{boot.ci}} with type="basic".
}
