% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_test.R
\name{generate_vulcano_data}
\alias{generate_vulcano_data}
\title{generate_vulcano_data}
\usage{
generate_vulcano_data(c_table, nf_b_table)
}
\arguments{
\item{c_table}{}
}
\value{

}
\description{
generate_vulcano_data
}
