% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EIC_table.R
\name{get_EIMatches_BM_NPPpeaks}
\alias{get_EIMatches_BM_NPPpeaks}
\title{get_EIMatches_BM_NPPpeaks}
\usage{
get_EIMatches_BM_NPPpeaks(rt, int, Min.PpP)
}
\arguments{
\item{rt}{rt}

\item{int}{int}

\item{Min.PpP}{Min.PpP}
}
\description{
get_EIMatches_BM_NPPpeaks
}
\keyword{internal}
