% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_funktionen_fuer_alignment.R
\name{get_main_UT_groups}
\alias{get_main_UT_groups}
\title{get_main_UT_groups}
\usage{
get_main_UT_groups(DT)
}
\arguments{
\item{DT}{}
}
\value{

}
\description{
Find most occuring feature ID within matched peaks of a benchmark feature
}
