% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_peak_vars.R
\name{Get_peak_vars}
\alias{Get_peak_vars}
\title{Get_peak_vars}
\usage{
Get_peak_vars(
  l.peaks,
  EIC.dt,
  CompCol_xic,
  l.peaks.mz_list,
  iso.run,
  adduct.run,
  manual_bound
)
}
\arguments{
\item{l.peaks}{l.peaks}

\item{EIC.dt}{EIC.dt}

\item{CompCol_xic}{CompCol_xic}

\item{l.peaks.mz_list}{l.peaks.mz_list}

\item{iso.run}{iso.run}

\item{adduct.run}{adduct.run}

\item{manual_bound}{manual_bound}
}
\description{
Get_peak_vars
}
\keyword{internal}
