% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pot_peak_ranges.R
\name{get_pot_peak_ranges2}
\alias{get_pot_peak_ranges2}
\title{get_pot_peak_ranges2}
\usage{
get_pot_peak_ranges2(int, Min.PpP = 10, peak.spotting.factor = 0.01)
}
\arguments{
\item{int}{int}

\item{Min.PpP}{Min.PpP}

\item{peak.spotting.factor}{peak.spotting.factor}
}
\description{
get_pot_peak_ranges2
}
\keyword{internal}
