% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEICfromROI_par.R
\name{get_ROIs}
\alias{get_ROIs}
\title{get_ROIs}
\usage{
get_ROIs(
  files,
  Target.table,
  minCentroids = 4,
  AccurateMZtol = 5,
  PrecisionMZtol = 5,
  plan = "multiprocess"
)
}
\arguments{
\item{files}{vector containing all mzML file paths}

\item{Target.table}{output of function \code{\link{get_mz_table}}}

\item{minCentroids}{minimum number of consecutive scans > 0 for a ROI to be picked up}

\item{AccurateMZtol}{mass accuracy (systematic error tolerance) in +/- ppm; this value is used to recognize detected ROIs as the expected mz values calculated in \code{\link{get_mz_table}}}

\item{PrecisionMZtol}{mass precision (random error tolerance) in +/- ppm; this value is used as for setting the maximum spread of scans within one ROI (equ. to "dev" argument in  xcms:::findmzROI)}

\item{plan}{see \code{\link{plan}}}
}
\value{
data.table object with information on ROIs for each row in Target.table. additional columns from Target.table are retained
}
\description{
Takes a the output of \code{\link{get_mz_table}} and adds columns with information on regions of interest (ROI).
}
