% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMassTraceList.R
\name{getMZtable}
\alias{getMZtable}
\title{getMZtable}
\usage{
getMZtable(
  DT,
  instrumentRes,
  RelInt_threshold = 0.05,
  stick_method = "intensoid",
  adducts,
  isotopes
)
}
\arguments{
\item{DT}{data.table with columns "molecule", "SumForm_c" and "adduct_c" (see details)}

\item{instrumentRes}{data frame with mz vs resolution dependence (see \code{\link{resolution_list}}) e.g. "resolution_list$`OTFusion,QExactiveHF_120000@200`"}

\item{RelInt_threshold}{relative abundance of the lowest isotopologue to be considered}

\item{stick_method}{method that should be used to calculate discrete m/z values from calculated profile pattern e.g. "intensoid" (see \code{\link{vdetect}})}

\item{adducts}{data frame with adducts (see \code{\link{adducts}})}

\item{isotopes}{data frame of isotopes (see \code{\link{isotopes}})}
}
\value{
data.table with columns "molecule", "formula", "adduct", "charge", "m/z" and "abundance"
}
\description{
This is basically a wrapper-function around the enviPat package. It takes a table with columns "molecule", "SumForm_c" and
"adduct_c" and calculates theoretical m/z values and abundances for isotopologues down to a user-defined abundance threshold.
}
\details{
Make sure that molecular formulas in column "SumForm_c" only contain valid molecular formulas as described in \code{\link{check_chemform}}. Otherwise function, might

never finish! Additional columns in DT will be retained in the output of the function. However, the column names "adduct", "isoabb", "formula", "charge" and "mz" are reserved.
}
