% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSharpness.R
\name{GetSharpness}
\alias{GetSharpness}
\title{GetSharpness}
\usage{
GetSharpness(int)
}
\arguments{
\item{int}{int}
}
\description{
As described in:
Zhang,W. and Zhao,P.X. (2014) Quality evaluation of extracted ion chromatograms and chromatographic peaks in
liquid chromatography/mass spectrometry-based metabolomics data. BMC Bioinformatics, 15, 1–13.
}
\keyword{internal}
