% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetZigzagIDX.R
\name{GetZigzagIDX}
\alias{GetZigzagIDX}
\title{GetZigzagIDX}
\usage{
GetZigzagIDX(int, height)
}
\arguments{
\item{int}{int}

\item{height}{height}
}
\description{
As described in:
Zhang,W. and Zhao,P.X. (2014) Quality evaluation of extracted ion chromatograms and chromatographic peaks in
liquid chromatography/mass spectrometry-based metabolomics data. BMC Bioinformatics, 15, 1–13.
}
\keyword{internal}
