% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_functions.R
\name{import_grouped_centwave}
\alias{import_grouped_centwave}
\title{import_grouped_centwave}
\usage{
import_grouped_centwave(file_path, options_table)
}
\arguments{
\item{options_table}{}
}
\description{
import_grouped_centwave
}
