% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_msdial.R
\name{import_grouped_msdial}
\alias{import_grouped_msdial}
\title{import_grouped_msdial}
\usage{
import_grouped_msdial(file_path, options_table)
}
\arguments{
\item{options_table}{}
}
\description{
import_grouped_msdial
}
