% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_openms.R
\name{import_grouped_openms}
\alias{import_grouped_openms}
\title{import_grouped_openms}
\usage{
import_grouped_openms(file_path, options_table)
}
\arguments{
\item{file_path}{file_path}

\item{options_table}{options_table}
}
\description{
import_grouped_openms
}
\keyword{internal}
