% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_slaw.R
\name{import_grouped_slaw}
\alias{import_grouped_slaw}
\title{import_grouped_slaw}
\usage{
import_grouped_slaw(file, options_dt)
}
\arguments{
\item{file}{file}

\item{options_dt}{options_dt}
}
\description{
import_grouped_slaw
}
\keyword{internal}
