% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_xcms.R
\name{import_grouped_xcms}
\alias{import_grouped_xcms}
\title{import_grouped_xcms}
\usage{
import_grouped_xcms(file, options_dt)
}
\arguments{
\item{file}{file}

\item{options_dt}{options_dt}
}
\description{
import_grouped_xcms
}
\keyword{internal}
