% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_compd.R
\name{import_ungrouped_cd}
\alias{import_ungrouped_cd}
\title{import_ungrouped_cd}
\usage{
import_ungrouped_cd(file_path, options_table)
}
\arguments{
\item{file_path}{file_path}

\item{options_table}{options_table}
}
\description{
import_ungrouped_cd
}
\keyword{internal}
