% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ungrouped_cd.R
\name{import_ungrouped_cd_depr}
\alias{import_ungrouped_cd_depr}
\title{import_ungrouped_cd_depr}
\usage{
import_ungrouped_cd_depr(file_path, options_table)
}
\arguments{
\item{options_table}{}
}
\value{

}
\description{
import_ungrouped_cd_depr
}
