% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_elmaven.R
\name{import_ungrouped_elmaven}
\alias{import_ungrouped_elmaven}
\title{import_ungrouped_elmaven}
\usage{
import_ungrouped_elmaven(file, options_dt)
}
\arguments{
\item{options_dt}{}
}
\value{

}
\description{
import_ungrouped_elmaven
}
