% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_Metaboanalyst.R
\name{import_ungrouped_Metaboanalyst}
\alias{import_ungrouped_Metaboanalyst}
\title{import_ungrouped_Metaboanalyst}
\usage{
import_ungrouped_Metaboanalyst(file, options_dt)
}
\arguments{
\item{file}{file}

\item{options_dt}{options_dt}
}
\description{
import_ungrouped_Metaboanalyst
}
\keyword{internal}
