% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_mzMine.R
\name{import_ungrouped_mzmine}
\alias{import_ungrouped_mzmine}
\title{import_ungrouped_mzmine}
\usage{
import_ungrouped_mzmine(folder_path, options_table)
}
\arguments{
\item{folder_path}{folder_path}

\item{options_table}{options_table}
}
\description{
import_ungrouped_mzmine
}
\keyword{internal}
