% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_openms.R
\name{import_ungrouped_openms}
\alias{import_ungrouped_openms}
\title{import_ungrouped_openms}
\usage{
import_ungrouped_openms(file_list, options_dt)
}
\arguments{
\item{options_dt}{}
}
\value{

}
\description{
import_ungrouped_openms
}
