% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_slaw.R
\name{import_ungrouped_slaw}
\alias{import_ungrouped_slaw}
\title{import_ungrouped_slaw}
\usage{
import_ungrouped_slaw(file_list, options_dt)
}
\arguments{
\item{file_list}{file_list}

\item{options_dt}{options_dt}
}
\description{
import_ungrouped_slaw
}
\keyword{internal}
