% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_xcms.R
\name{import_ungrouped_xcms}
\alias{import_ungrouped_xcms}
\title{import_ungrouped_xcms}
\usage{
import_ungrouped_xcms(file_path, options_table)
}
\arguments{
\item{options_table}{}
}
\description{
import_ungrouped_xcms
}
