% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyPeakIntegration_par2.R
\name{LazyPeakIntegration_par}
\alias{LazyPeakIntegration_par}
\title{LazyPeakIntegration_par}
\usage{
LazyPeakIntegration_par(files, Grps, CompCol_all, Min.PpP = 5, ppm,
  Nr.of.acceptedMissingScans, PredictedRT = FALSE,
  Cluster = makeCluster(detectCores() - 2))
}
\arguments{
\item{files}{vector containing mzML file paths}

\item{Grps}{dataframe with two columns; one with file name, one with sample group}

\item{Min.PpP}{minimum number of points per peak for a peak candidate}

\item{ppm}{maxmimum m/z deviation from the theortical mass}

\item{Nr.of.acceptedMissingScans}{maxmium number of consecutive missing data points (see \code{\link{GetConsecScanBorders}})}

\item{Cluster}{number of cores to be used in parallel}

\item{CompCol}{dataframe containing names, sumformulas, m/z values and integration boundaries}
}
\value{
The output from \code{\link{rbindlist}}
}
\description{
Takes a list of named m/z values and rough integration bondaries in the retention time dimension. Chromatograms implied are
extracted from a list of mzML files. A peak is picked from each chromatogram. Multiple variables are extracted from each peak...
}
\examples{
\dontrun{
peak_candidadtes <- LazyPeakIntegration(files, Grps_all, CompColred, 10, 5, 2)
}

}
