% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limit_Target_list.R
\name{Limit_Target_list}
\alias{Limit_Target_list}
\title{Limit_Target_list}
\usage{
Limit_Target_list(
  CompCol,
  CondPeaks,
  iso.run,
  adduct.run,
  file,
  Min.PointsperPeak
)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Limit_Target_list
}
