% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_find_main_feature.R
\name{match_features_by_iso}
\alias{match_features_by_iso}
\title{match_features_by_iso}
\usage{
match_features_by_iso(dt, max_possible_found_peaks, nf_g_table, factor)
}
\arguments{
\item{factor}{}
}
\description{
match_features_by_iso
}
