% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features_to_benchmark.R
\name{match_features_to_benchmark}
\alias{match_features_to_benchmark}
\title{match_features_to_benchmark}
\usage{
match_features_to_benchmark(
  g_table,
  b_table,
  Matches_BM_NPPpeaks,
  Unmatched_BM_NPPpeaks
)
}
\arguments{
\item{g_table}{g_table}

\item{b_table}{b_table}

\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{Unmatched_BM_NPPpeaks}{Unmatched_BM_NPPpeaks}
}
\description{
match_features_to_benchmark
}
\keyword{internal}
