% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_NPPpeaks_to_NPPfeatures.R
\name{match_NPPpeaks_to_NPPfeatures}
\alias{match_NPPpeaks_to_NPPfeatures}
\title{match_NPPpeaks_to_NPPfeatures}
\usage{
match_NPPpeaks_to_NPPfeatures(
  Matches_BM_NPPpeaks,
  SplittedMatches_BM_NPPpeaks,
  g_table
)
}
\arguments{
\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{SplittedMatches_BM_NPPpeaks}{SplittedMatches_BM_NPPpeaks}

\item{g_table}{g_table}
}
\description{
match_NPPpeaks_to_NPPfeatures
}
\keyword{internal}
