% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_peaks_to_benchmark.R
\name{match_peaks_to_benchmark}
\alias{match_peaks_to_benchmark}
\title{match_peaks_to_benchmark}
\usage{
match_peaks_to_benchmark(b_table, ug_table)
}
\arguments{
\item{b_table}{b_table}

\item{ug_table}{ug_table}
}
\description{
match_peaks_to_benchmark
}
\keyword{internal}
