% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_peaks_to_benchmark_split.R
\name{match_peaks_to_benchmark_split}
\alias{match_peaks_to_benchmark_split}
\title{match_peaks_to_benchmark_split}
\usage{
match_peaks_to_benchmark_split(b_table, ug_table)
}
\arguments{
\item{b_table}{b_table}

\item{ug_table}{ug_table}
}
\description{
match_peaks_to_benchmark_split
}
\keyword{internal}
