% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_per_molecule.R
\name{metrics_per_molecule}
\alias{metrics_per_molecule}
\title{metrics_per_molecule}
\usage{
metrics_per_molecule(
  Matches_BM_NPPpeaks,
  Unmatched_BM_NPPpeaks,
  Matches_BM_NPPpeaks_NPPfeatures,
  IT_ratio_biases,
  SplittedMatches_BM_NPPpeaks,
  MissingPeak_classification,
  AlignmentErrors_per_moleculeAndAdduct
)
}
\arguments{
\item{Matches_BM_NPPpeaks}{Matches_BM_NPPpeaks}

\item{Unmatched_BM_NPPpeaks}{Unmatched_BM_NPPpeaks}

\item{Matches_BM_NPPpeaks_NPPfeatures}{Matches_BM_NPPpeaks_NPPfeatures}

\item{IT_ratio_biases}{IT_ratio_biases}

\item{SplittedMatches_BM_NPPpeaks}{SplittedMatches_BM_NPPpeaks}

\item{MissingPeak_classification}{MissingPeak_classification}

\item{AlignmentErrors_per_moleculeAndAdduct}{AlignmentErrors_per_moleculeAndAdduct}
}
\description{
metrics_per_molecule
}
\keyword{internal}
