% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_pick_algo.R
\name{pick_algorithm}
\alias{pick_algorithm}
\title{pick_algorithm}
\usage{
pick_algorithm(ug_table_path, g_table_path, options_table, algo)
}
\arguments{
\item{ug_table_path}{path to unaligned table(s)}

\item{g_table_path}{path to aligned table}

\item{options_table}{output from \code{\link{import_benchmark}}}

\item{algo}{tool output format of ug_table and g_table. can be XCMS, El-Maven, OpenMS, msDial, CompoundDiscoverer or mzMine. Outputs from different tools can also be used as long as they are reformatted to one of those types.}
}
\value{
returns unaligned and aligned outputs from non-targeted tool in a format readable via \code{\link{compare_peaks}}
}
\description{
pick_algorithm
}
