% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_pick_main_feature.R
\name{pick_main_feature_sd}
\alias{pick_main_feature_sd}
\title{pick_main_feature_sd}
\usage{
pick_main_feature_sd(dt)
}
\arguments{
\item{dt}{dt}
}
\description{
pick_main_feature_sd
}
\keyword{internal}
