% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_heatmap.R
\name{plot_bench_heatmap}
\alias{plot_bench_heatmap}
\title{plot_bench_heatmap}
\usage{
plot_bench_heatmap(benchmark_data)
}
\arguments{
\item{benchmark_data}{}
}
\value{

}
\description{
plot_bench_heatmap
}
