% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_histo.R
\name{plot_bench_histo}
\alias{plot_bench_histo}
\title{plot_bench_histo}
\usage{
plot_bench_histo(
  benchmark_data,
  var,
  choice_vector_bench,
  color = "blue",
  post_comp = FALSE,
  rm_NF_legend = FALSE
)
}
\arguments{
\item{benchmark_data}{output from \code{\link{findBenchPeaks}}}

\item{var}{variable name to be plotted}

\item{choice_vector_bench}{named vector including variable to be plotted as element}
}
\value{
plotly object
}
\description{
plot_bench_histo
}
