% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_overview.R
\name{plot_bench_overview}
\alias{plot_bench_overview}
\title{plot_bench_overview}
\usage{
plot_bench_overview(benchmark_data, x, y, colb, choice_vector_bench)
}
\arguments{
\item{choice_vector_bench}{}
}
\description{
plot_bench_overview
}
