% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_peak_overview.R
\name{plot_bench_peak_overview}
\alias{plot_bench_peak_overview}
\title{plot_bench_peak_overview}
\usage{
plot_bench_peak_overview(benchmark_data, molecule, adduct, ia)
}
\arguments{
\item{ia}{}
}
\description{
plot_bench_peak_overview
}
