% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_dist_of_found_peaks.R
\name{plot_comp_dist_of_found_peaks}
\alias{plot_comp_dist_of_found_peaks}
\title{plot_comp_dist_of_found_peaks}
\usage{
plot_comp_dist_of_found_peaks(
  comparison_data,
  var,
  choice_vector_comp,
  post_alignment = FALSE
)
}
\arguments{
\item{comparison_data}{output from \code{\link{compare_peaks}}}

\item{var}{variable (column name) to be plotted}

\item{choice_vector_comp}{named vector containing variable to be plotted as element}

\item{post_alignment}{TRUE/FALSE should data be plotted from before or after alignment.}
}
\value{
plotly object
}
\description{
plot_comp_dist_of_found_peaks
}
