% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_iso_pred_error.R
\name{plot_comp_iso_pred_error}
\alias{plot_comp_iso_pred_error}
\title{plot_comp_iso_pred_error}
\usage{
plot_comp_iso_pred_error(
  comparison_data,
  post_alignment = FALSE,
  BMvsPPvsAl = TRUE
)
}
\arguments{
\item{comparison_data}{output from \code{\link{compare_peaks}}}

\item{post_alignment}{TRUE/FALSE should NT data from before or after alignment be plotted}

\item{BMvsPPvsAl}{TRUE/FALSE should argument post_alignment be ignored in order to plot both in one plot}
}
\value{
plotly object
}
\description{
plot_comp_iso_pred_error
}
