% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_missing_value_hm.R
\name{plot_comp_missing_value_hm}
\alias{plot_comp_missing_value_hm}
\title{plot_comp_missing_value_hm}
\usage{
plot_comp_missing_value_hm(
  comparison_data,
  post_alignment = FALSE,
  disable_plot = FALSE
)
}
\arguments{
\item{comparison_data}{output from \code{\link{compare_peaks}}}

\item{post_alignment}{TRUE/FALSE should NT data from before or after alignment be plotted}

\item{disable_plot}{if TRUE plot is not generated (for shiny app due to long loading times)}
}
\value{
plotly object
}
\description{
plot_comp_missing_value_hm
}
