% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_peak_overview.R
\name{plot_comp_peak_overview}
\alias{plot_comp_peak_overview}
\title{plot_comp_peak_overview}
\usage{
plot_comp_peak_overview(comparison_data, mol_c, add_c, ia_c)
}
\arguments{
\item{comparison_data}{output from \code{\link{compare_peaks}}}

\item{mol_c}{molecule}

\item{add_c}{adduct}

\item{ia_c}{isotopic abundance rounded to 2 digits}
}
\value{
plotly object
}
\description{
plot_comp_peak_overview
}
