% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_scatter_plot.R
\name{plot_comp_scatter_plot}
\alias{plot_comp_scatter_plot}
\title{plot_comp_scatter_plot}
\usage{
plot_comp_scatter_plot(
  comparison_data,
  x,
  y,
  col,
  choice_vector_comp,
  post_alignment = FALSE
)
}
\arguments{
\item{comparison_data}{output from \code{\link{compare_peaks}}}

\item{x}{variable (column name) to be plotted on x axis}

\item{y}{variable (column name) to be plotted on y axis}

\item{col}{variable (column name) to be colored by}

\item{choice_vector_comp}{named vector including all variables used as elements}

\item{post_alignment}{TRUE/FALSE should NT data from before or after alignment be plotted}
}
\value{
plotly object
}
\description{
plot_comp_scatter_plot
}
