% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_IR_peaks}
\alias{plot_IR_peaks}
\title{plot_IR_peaks}
\usage{
plot_IR_peaks(PC, plotly_key)
}
\arguments{
\item{PC}{PC}

\item{plotly_key}{plotly_key}
}
\description{
plot_IR_peaks
}
\keyword{internal}
