% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_Peak}
\alias{plot_Peak}
\title{plot_Peak}
\usage{
plot_Peak(PC, IndexNumber)
}
\arguments{
\item{PC}{output from \code{\link{find_bench_peaks}}}

\item{IndexNumber}{IDX number of peak to be plotted}
}
\value{
plotly object
}
\description{
plot_Peak
}
