% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plots.R
\name{plot_sunburst_peakQuality}
\alias{plot_sunburst_peakQuality}
\title{plot_sunburst_peakQuality}
\usage{
plot_sunburst_peakQuality(result_txt, comparison_object)
}
\arguments{
\item{comparison_object}{}
}
\value{

}
\description{
plot_sunburst_peakQuality
}
