% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vs_prediction.R
\name{plot_vs_prediction}
\alias{plot_vs_prediction}
\title{plot_vs_prediction}
\usage{
plot_vs_prediction(PC, x = ExpectedArea, y, col_by = molecule)
}
\arguments{
\item{y}{}
}
\description{
plot_vs_prediction
}
