% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{plotly_click_wo_warnings}
\alias{plotly_click_wo_warnings}
\title{plotly_click_wo_warnings}
\usage{
plotly_click_wo_warnings(sc)
}
\arguments{
\item{sc}{sc}
}
\value{
dataframe; plotly event
}
\description{
plotly_click_wo_warnings
}
\keyword{internal}
