% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_functions.R
\name{rename_columns_from_options}
\alias{rename_columns_from_options}
\title{rename_columns_from_options}
\usage{
rename_columns_from_options(dt, options_dt, old_columns, new_columns)
}
\arguments{
\item{dt}{dt}

\item{new_columns}{new_columns}

\item{options_table}{options_table}

\item{old_column}{old_column}
}
\description{
rename_columns_from_options
}
\keyword{internal}
