% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_functions.R
\name{SkylinePeakBoundaries}
\alias{SkylinePeakBoundaries}
\title{SkylinePeakBoundaries}
\usage{
SkylinePeakBoundaries(BM)
}
\arguments{
\item{BM}{output of \code{\link{align_PC}}}
}
\value{
Skyline peak boundaries
}
\description{
Takes a the output of \code{\link{align_PC}} and generates a Peak boundaries which can then be imported to Skyline via
Skyline -> File -> Import -> Peak Boundaries...
}
