% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{top_to_x}
\alias{top_to_x}
\title{top_to_x}
\usage{
top_to_x(number, x = 0)
}
\arguments{
\item{number}{numeric(1)}

\item{x}{roof to this number}
}
\value{
numeric
}
\description{
top_to_x
}
\keyword{internal}
