% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_error_plot.R
\name{Alignment_error_plot}
\alias{Alignment_error_plot}
\title{Alignment_error_plot}
\usage{
Alignment_error_plot(comparison_data, mol, add)
}
\arguments{
\item{comparison_data}{}
}
\value{

}
\description{
Alignment_error_plot
}
