% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_funktionen für alignment test_2.R
\name{count_alignment_errors}
\alias{count_alignment_errors}
\title{count_alignment_errors}
\usage{
count_alignment_errors(DT, main_UTgroups, method = "self-critical")
}
\arguments{
\item{method}{}
}
\description{
count alignment errors occuring within a compound
}
