% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_double_peaks.R
\name{detect_double_peaks2}
\alias{detect_double_peaks2}
\title{detect_double_peaks2}
\usage{
detect_double_peaks2(
  pot.doubleP.v,
  Min.PpP = 10,
  l = 1,
  r = length(pot.doubleP.v),
  Min.Res = 70
)
}
\arguments{
\item{pot.doubleP.v}{pot.doubleP.v}

\item{Min.PpP}{Min.PpP}

\item{l}{l}

\item{r}{r}

\item{Min.Res}{Min.Res}
}
\description{
detect_double_peaks2
}
\keyword{internal}
