% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LazyPeakIntegration_peaks.R
\name{findBenchPeaks}
\alias{findBenchPeaks}
\title{findBenchPeaks}
\usage{
findBenchPeaks(files, Grps, CompCol_all, Min.PointsperPeak = 10,
  peak.spotting.factor = 0.001, Integration_baseL_factor = 0.1,
  Min.Res = 60, plan = "multiprocess", Min.cor.w.main_adduct = 0.8,
  Min.cor.w.M0 = 0.8, Min.iso.count = 2,
  return_unsuc_searches = FALSE)
}
\arguments{
\item{Min.Res}{}
}
\description{
findBenchPeaks
}
