% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_generate_results_text.R
\name{generate_results_text}
\alias{generate_results_text}
\title{generate_results_text}
\usage{
generate_results_text(comparison_data)
}
\arguments{
\item{comparison}{}
}
\description{
generate_results_text
}
