% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_funktionen für alignment test_2.R
\name{get_main_UT_groups}
\alias{get_main_UT_groups}
\title{get_main_UT_groups}
\usage{
get_main_UT_groups(DT)
}
\arguments{
\item{DT}{}
}
\value{

}
\description{
get_main_UT_groups
}
