% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_peak_vars.R
\name{Get_peak_vars}
\alias{Get_peak_vars}
\title{Get_peak_vars}
\usage{
Get_peak_vars(
  l.peaks,
  EIC.dt,
  CompCol_xic,
  l.peaks.mz_list,
  iso.run,
  adduct.run,
  manual_bound
)
}
\arguments{
\item{l.peaks.mz_list}{}
}
\value{

}
\description{
Get_peak_vars
}
