% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_benchmark.R
\name{import_benchmark}
\alias{import_benchmark}
\title{import_benchmark}
\usage{
import_benchmark(file, options_path, from_csv = TRUE, algo)
}
\arguments{
\item{from_csv}{}
}
\description{
import_benchmark
}
