% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_mzMine.R
\name{import_grouped_mzmine}
\alias{import_grouped_mzmine}
\title{import_grouped_mzmine}
\usage{
import_grouped_mzmine(file_path, options_table)
}
\arguments{
\item{options_table}{}
}
\value{

}
\description{
import_grouped_mzmine
}
