% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_import_compd.R, R/import_ungrouped_cd.R
\name{import_ungrouped_cd}
\alias{import_ungrouped_cd}
\title{import_ungrouped_cd}
\usage{
import_ungrouped_cd(file_path, options_table)

import_ungrouped_cd(file_path, options_table)
}
\arguments{
\item{options_table}{}

\item{options_table}{}
}
\description{
import_ungrouped_cd

import_ungrouped_cd
}
