% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_pick_algo.R
\name{pick_algorithm}
\alias{pick_algorithm}
\title{pick_algorithm}
\usage{
pick_algorithm(ug_table_path, g_table_path, options_table, algo)
}
\arguments{
\item{algo}{}
}
\value{

}
\description{
pick_algorithm
}
