% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_histo.R
\name{plot_bench_histo}
\alias{plot_bench_histo}
\title{plot_bench_histo}
\usage{
plot_bench_histo(
  benchmark_data,
  var,
  choice_vector_bench,
  color = "blue",
  post_comp = FALSE,
  rm_NF_legend = FALSE
)
}
\arguments{
\item{choice_vector_bench}{}
}
\value{

}
\description{
plot_bench_histo
}
