% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bench_peak_overview.R
\name{plot_bench_peak_overview}
\alias{plot_bench_peak_overview}
\title{plot_bench_peak_overview}
\usage{
plot_bench_peak_overview(benchmark_data, molecule, adduct, ia)
}
\arguments{
\item{benchmark_data}{output from \code{\link{findBenchPeaks}}}

\item{molecule}{molecule}

\item{adduct}{adduct}

\item{ia}{isotopic abundance rounded to 2 digits}
}
\value{
plotly object
}
\description{
plot_bench_peak_overview
}
