% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_dist_of_found_peaks.R
\name{plot_comp_dist_of_found_peaks}
\alias{plot_comp_dist_of_found_peaks}
\title{plot_comp_dist_of_found_peaks}
\usage{
plot_comp_dist_of_found_peaks(
  comparison_data,
  var,
  choice_vector_comp,
  post_alignment = FALSE
)
}
\arguments{
\item{post_alignment}{}
}
\value{

}
\description{
plot_comp_dist_of_found_peaks
}
