% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_iso_pred_error.R
\name{plot_comp_iso_pred_error}
\alias{plot_comp_iso_pred_error}
\title{plot_comp_iso_pred_error}
\usage{
plot_comp_iso_pred_error(comparison_data)
}
\arguments{
\item{comparison_data}{}
}
\description{
plot_comp_iso_pred_error
}
