% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_missing_value_hm.R
\name{plot_comp_missing_value_hm}
\alias{plot_comp_missing_value_hm}
\title{plot_comp_missing_value_hm}
\usage{
plot_comp_missing_value_hm(comparison_data)
}
\arguments{
\item{comparison_data}{}
}
\value{

}
\description{
plot_comp_missing_value_hm
}
