% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_scatter_plot.R
\name{plot_comp_scatter_plot}
\alias{plot_comp_scatter_plot}
\title{plot_comp_scatter_plot}
\usage{
plot_comp_scatter_plot(comparison_data, x, y, choice_vector_comp)
}
\arguments{
\item{plot_input_y}{}
}
\description{
plot_comp_scatter_plot
}
