% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_Peak}
\alias{plot_Peak}
\title{plot_Peak}
\usage{
plot_Peak(PC, IndexNumber)
}
\arguments{
\item{PC}{PeakCandidates tabel from from which peak should be plotted}

\item{IndexNumber}{IDX number of peak to be plotted}
}
\value{

}
\description{
plot_Peak
}
