% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_Peak_per_mol}
\alias{plot_Peak_per_mol}
\title{plot_Peak_per_mol}
\usage{
plot_Peak_per_mol(PC_object, mol, ia = 100, add = "M+H")
}
\arguments{
\item{PC_object}{output from \code{\link{compare_peaks}} or output from \code{\link{findBenchPeaks}}}

\item{mol}{molecule}

\item{ia}{isotopic abundance rounded to 2 digits}

\item{add}{adduct}
}
\value{
plotly object
}
\description{
plot_Peak_per_mol
}
