% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Peak.R
\name{plot_Peak_with_predicted_peak}
\alias{plot_Peak_with_predicted_peak}
\title{plot_Peak_with_predicted_peak}
\usage{
plot_Peak_with_predicted_peak(PC_object, IndexNumber)
}
\arguments{
\item{PC_object}{output from \code{\link{find_bench_peaks}}}

\item{IndexNumber}{IDX number of peak to be plotted}
}
\value{
plotly object
}
\description{
plot_Peak_with_predicted_peak
}
