% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plots.R
\name{plot_sunburst_alignment}
\alias{plot_sunburst_alignment}
\title{plot_sunburst_alignment}
\usage{
plot_sunburst_alignment(result_txt)
}
\arguments{
\item{result_txt}{}
}
\value{

}
\description{
plot_sunburst_alignment
}
