% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plots.R
\name{plot_sunburst_peaks}
\alias{plot_sunburst_peaks}
\title{plot_sunburst_peaks}
\usage{
plot_sunburst_peaks(result_txt, comparison_object)
}
\arguments{
\item{comparison_object}{}
}
\value{

}
\description{
plot_sunburst_peaks
}
