% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst_plots.R
\name{plot_sunburst_peaks}
\alias{plot_sunburst_peaks}
\title{plot_sunburst_peaks}
\usage{
plot_sunburst_peaks(result_txt, comparison_object)
}
\arguments{
\item{result_txt}{output from \code{\link{generate_results_text}}}

\item{comparison_object}{output from \code{\link{compare_peaks}}}
}
\value{
plotly object
}
\description{
plot_sunburst_peaks
}
