% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Iso.R
\name{predict_Iso}
\alias{predict_Iso}
\title{predict_Iso}
\usage{
predict_Iso(DTT, SampleIdentifier_col, Molecule_Adduct_col, IsoAbb_col,
  flag_extremes = FALSE)
}
\arguments{
\item{flag_extremes}{}
}
\description{
predict_Iso
}
