% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_functions.R
\name{remove_identical_peaks}
\alias{remove_identical_peaks}
\title{remove_identical_peaks}
\usage{
remove_identical_peaks(dt, grouped = FALSE)
}
\arguments{
\item{dt}{dt}

\item{incl_height}{height}
}
\value{
dpl
}
\description{
remove_identical_peaks
}
\keyword{internal}
