% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_functions.R
\name{rename_columns_from_options}
\alias{rename_columns_from_options}
\title{rename_columns_from_options}
\usage{
rename_columns_from_options(dt, options_dt, old_columns, new_columns)
}
\arguments{
\item{new_columns}{}
}
\value{

}
\description{
rename_columns_from_options
}
