% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_functions.R
\name{SkylinePeakBoundaries}
\alias{SkylinePeakBoundaries}
\title{SkylinePeakBoundaries}
\usage{
SkylinePeakBoundaries(BM)
}
\arguments{
\item{BM}{}
}
\description{
SkylinePeakBoundaries
}
\examples{

}
