% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_functions.R
\name{SkylinePeakBoundaries}
\alias{SkylinePeakBoundaries}
\title{SkylinePeakBoundaries}
\usage{
SkylinePeakBoundaries(BM)
}
\arguments{
\item{BM}{output of \code{\link{find_bench_peaks}}}
}
\value{
Skyline peak boundaries
}
\description{
Takes the output of \code{\link{find_bench_peaks}} and generates a Skyline peak-boundaries file (automatically exported to working directory) which can then be imported to Skyline via
Skyline -> File -> Import -> Peak Boundaries...
}
