% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_functions.R
\name{SkylineTransitionList}
\alias{SkylineTransitionList}
\title{SkylineTransitionList}
\usage{
SkylineTransitionList(BM)
}
\arguments{
\item{BM}{}
}
\description{
SkylineTransitionList
}
\examples{

}
