% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_functions.R
\name{SkylineTransitionList}
\alias{SkylineTransitionList}
\title{SkylineTransitionList}
\usage{
SkylineTransitionList(BM)
}
\arguments{
\item{BM}{output of \code{\link{find_bench_peaks}}}
}
\value{
Skyline Transtion List
}
\description{
Takes a the output of \code{\link{find_bench_peaks}} and generates a Skyline Transition list which can then be imported to Skyline via
Skyline -> Settings -> Transition Settings -> Full-Scan -> Mass Accuracy
}
